refactor: team code generator 분리
